Fixed infinite loop in Guard.
For some, unknown reason, Guard was triggering on-modification
callbacks whenever anonymized.sqlite3 was touched, even though
it was only accessed, not modified, i.e. the atime was updated
but not the ctime. This triggered an infinite loop in Guard
that manifested after the first change or when target "all" was
run by pressing Enter. Root cause unknown.